singular-agents/academic-agents/requirements.txt
# ============================================
# Core Framework
# ============================================
Flask
flask-cors
python-dotenv
redis

# ============================================
# PDF and Image Processing
# ============================================
pdfplumber
PyPDF2
Pillow

# ============================================
# LangChain + Gemini
# ============================================
langchain-core
langchain-google-genai
google-generativeai

# (Optional – only if you actually use Groq)
langchain-groq

# ============================================
# HTTP / Async
# ============================================
requests
aiohttp
httpx

# ============================================
# Data Processing
# ============================================
numpy
pandas

# ============================================
# Dev & Testing
# ============================================
black
pytest
pytest-asyncio
